move <Resource/> into its own folder
