+++ src/components/Resource/Resource.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import Resource from './Resource';
-import { foodResource } from '../DataMocks';
+import { foodResource } from '../../DataMocks';
 
 describe('<Resource/>', () => {
   const mockSingleResourceId = '-KS4rbQziVbJM5tNdEQL';
