+++ src/components/Resource/Resource.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { TResource } from '../types';
+import { TResource } from '../../types';
+
 interface Props {
   id: string;
   resource: TResource;
